Revert "new resource and data source : azurerm_oracle_cloud_vm_cluster (#27176)"
This reverts commit cb8d847dc2e10ce7987234cf61d64dd59ec1e511.
