+++ internal/services/oracledatabase/exadata_infrastructure_resource.go
@@ -1,6 +1,6 @@
 // Copyright © 2024, Oracle and/or its affiliates. All rights reserved
 
-package oracle
+package oracledatabase
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	"github.com/hashicorp/go-azure-helpers/resourcemanager/zones"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/oracledatabase/2024-06-01/cloudexadatainfrastructures"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/sdk"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/oracle/validate"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/services/oracledatabase/validate"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 	"github.com/hashicorp/terraform-provider-azurerm/utils"
 )
@@ -179,7 +179,7 @@ func (r ExadataInfraResource) Create() sdk.ResourceFunc {
 	return sdk.ResourceFunc{
 		Timeout: 120 * time.Minute,
 		Func: func(ctx context.Context, metadata sdk.ResourceMetaData) error {
-			client := metadata.Client.Oracle.OracleClient.CloudExadataInfrastructures
+			client := metadata.Client.OracleDatabase.OracleDatabaseClient.CloudExadataInfrastructures
 			subscriptionId := metadata.Client.Account.SubscriptionId
 
 			var model ExadataInfraResourceModel
@@ -240,7 +240,7 @@ func (r ExadataInfraResource) Update() sdk.ResourceFunc {
 		Timeout: 30 * time.Minute,
 		Func: func(ctx context.Context, metadata sdk.ResourceMetaData) error {
 
-			client := metadata.Client.Oracle.OracleClient.CloudExadataInfrastructures
+			client := metadata.Client.OracleDatabase.OracleDatabaseClient.CloudExadataInfrastructures
 			id, err := cloudexadatainfrastructures.ParseCloudExadataInfrastructureID(metadata.ResourceData.Id())
 			if err != nil {
 				return err
@@ -286,7 +286,7 @@ func (ExadataInfraResource) Read() sdk.ResourceFunc {
 				return err
 			}
 
-			client := metadata.Client.Oracle.OracleClient.CloudExadataInfrastructures
+			client := metadata.Client.OracleDatabase.OracleDatabaseClient.CloudExadataInfrastructures
 			result, err := client.Get(ctx, *id)
 			if err != nil {
 				if response.WasNotFound(result.HttpResponse) {
@@ -323,7 +323,7 @@ func (ExadataInfraResource) Delete() sdk.ResourceFunc {
 	return sdk.ResourceFunc{
 		Timeout: 60 * time.Minute,
 		Func: func(ctx context.Context, metadata sdk.ResourceMetaData) error {
-			client := metadata.Client.Oracle.OracleClient.CloudExadataInfrastructures
+			client := metadata.Client.OracleDatabase.OracleDatabaseClient.CloudExadataInfrastructures
 
 			id, err := cloudexadatainfrastructures.ParseCloudExadataInfrastructureID(metadata.ResourceData.Id())
 			if err != nil {
