+++ internal/services/oracledatabase/exadata_infrastructure_resource_test.go
@@ -1,6 +1,6 @@
 // Copyright © 2024, Oracle and/or its affiliates. All rights reserved
 
-package oracle_test
+package oracledatabase_test
 
 import (
 	"context"
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/go-azure-sdk/resource-manager/oracledatabase/2024-06-01/cloudexadatainfrastructures"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/oracle"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/services/oracledatabase"
 
 	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance/check"
@@ -24,7 +24,7 @@ func (a ExadataInfraResource) Exists(ctx context.Context, client *clients.Client
 	if err != nil {
 		return nil, err
 	}
-	resp, err := client.Oracle.OracleClient.CloudExadataInfrastructures.Get(ctx, *id)
+	resp, err := client.OracleDatabase.OracleDatabaseClient.CloudExadataInfrastructures.Get(ctx, *id)
 	if err != nil {
 		return nil, fmt.Errorf("retrieving Exadata Infrastructure %s: %+v", id, err)
 	}
@@ -36,14 +36,14 @@ func (a ExadataInfraResource) Destroy(ctx context.Context, client *clients.Clien
 	if err != nil {
 		return nil, err
 	}
-	if _, err := client.Oracle.OracleClient.CloudExadataInfrastructures.Delete(ctx, *id); err != nil {
+	if _, err := client.OracleDatabase.OracleDatabaseClient.CloudExadataInfrastructures.Delete(ctx, *id); err != nil {
 		return nil, fmt.Errorf("deleting Exadata Infrastructure %s: %+v", id, err)
 	}
 	return utils.Bool(true), nil
 }
 
 func TestExaInfra_basic(t *testing.T) {
-	data := acceptance.BuildTestData(t, oracle.ExadataInfraResource{}.ResourceType(), "test")
+	data := acceptance.BuildTestData(t, oracledatabase.ExadataInfraResource{}.ResourceType(), "test")
 	r := ExadataInfraResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
@@ -57,7 +57,7 @@ func TestExaInfra_basic(t *testing.T) {
 }
 
 func TestExaInfra_allFields(t *testing.T) {
-	data := acceptance.BuildTestData(t, oracle.ExadataInfraResource{}.ResourceType(), "test")
+	data := acceptance.BuildTestData(t, oracledatabase.ExadataInfraResource{}.ResourceType(), "test")
 	r := ExadataInfraResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
@@ -71,7 +71,7 @@ func TestExaInfra_allFields(t *testing.T) {
 }
 
 func TestExaInfra_requiresImport(t *testing.T) {
-	data := acceptance.BuildTestData(t, oracle.ExadataInfraResource{}.ResourceType(), "test")
+	data := acceptance.BuildTestData(t, oracledatabase.ExadataInfraResource{}.ResourceType(), "test")
 	r := ExadataInfraResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
@@ -85,7 +85,7 @@ func TestExaInfra_requiresImport(t *testing.T) {
 }
 
 func TestExaInfra_update(t *testing.T) {
-	data := acceptance.BuildTestData(t, oracle.ExadataInfraResource{}.ResourceType(), "test")
+	data := acceptance.BuildTestData(t, oracledatabase.ExadataInfraResource{}.ResourceType(), "test")
 	r := ExadataInfraResource{}
 	data.ResourceTest(t, r, []acceptance.TestStep{
 		{
