+++ internal/services/oracledatabase/registration.go
@@ -1,6 +1,6 @@
 // Copyright © 2024, Oracle and/or its affiliates. All rights reserved
 
-package oracle
+package oracledatabase
 
 import (
 	"github.com/hashicorp/terraform-provider-azurerm/internal/sdk"
@@ -14,27 +14,24 @@ var (
 
 func (r Registration) DataSources() []sdk.DataSource {
 	return []sdk.DataSource{
-		CloudVmClusterDataSource{},
-		DBServersDataSource{},
 		ExadataInfraDataSource{},
 	}
 }
 
 func (r Registration) Resources() []sdk.Resource {
 	return []sdk.Resource{
-		CloudVmClusterResource{},
 		ExadataInfraResource{},
 	}
 }
 
 // Name is the name of this Service
 func (r Registration) Name() string {
-	return "Oracle"
+	return "Oracle Database"
 }
 
 // WebsiteCategories returns a list of categories which can be used for the sidebar
 func (r Registration) WebsiteCategories() []string {
 	return []string{
-		"Oracle",
+		"Oracle Database",
 	}
 }
