+++ internal/services/oracledatabase/validate/exadata_infrastructure.go
@@ -129,7 +129,7 @@ func WeeksOfMonth(i interface{}, k string) (warnings []string, errors []error) {
 func Preference(i interface{}, k string) (warnings []string, errors []error) {
 	v, ok := i.(string)
 	if !ok {
-		errors = append(errors, fmt.Errorf("expected type of %s to be string", k))
+		errors = append(errors, fmt.Errorf("expected type of %s to be list of strings", k))
 		return
 	}
 
@@ -145,7 +145,7 @@ func Preference(i interface{}, k string) (warnings []string, errors []error) {
 func PatchingMode(i interface{}, k string) (warnings []string, errors []error) {
 	v, ok := i.(string)
 	if !ok {
-		errors = append(errors, fmt.Errorf("expected type of %s to be string", k))
+		errors = append(errors, fmt.Errorf("expected type of %s to be list of strings", k))
 		return
 	}
