+++ website/allowed-subcategories
@@ -79,7 +79,7 @@ NetApp
 Network
 Network Function
 New Relic
-Oracle
+Oracle Database
 Orbital
 Palo Alto
 Policy
